python/requirements.txt
Jinja2==2.10.1
MarkupSafe==1.1.1
Werkzeug==0.15.2
click==7.0
flask==1.0.2
itsdangerous==1.1.0
python-dotenv==0.10.1
mysql-connector==2.2.9
mysql-connector-python==8.0.16
protobuf==3.7.1
six==1.12.0